requirements.txt
pandas>=2.2.0
streamlit>=1.40.0
plotly>=5.24.0
pdfplumber>=0.11.0
ofxparse>=0.21
python-dateutil>=2.9.0
pyyaml>=6.0
